pyproject.toml
[tool.poetry]
name = "buy_me_a_gift"
version = "1.0.0"
description = "Buy Me A Gift Challenge Project"
authors = ["abelayalew <abelayalew81@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.5"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
